add mpi version and move to integers

+++ tasks/gutyansky_a_matrix_column_sum/common/include/matrix.hpp
@@ -2,13 +2,14 @@
 
 #include <cstddef>
 #include <vector>
+#include <cmath>
 
 namespace gutyansky_a_matrix_column_sum {
 
 struct Matrix {
   size_t rows;
   size_t cols;
-  std::vector<double> data;
+  std::vector<int64_t> data;
 
   friend bool operator==(const Matrix &v_left, const Matrix &v_right) {
     return v_left.rows == v_right.rows && v_left.cols == v_right.cols && v_left.data == v_right.data;
